# Cache results of `kempner_with_data` and other related computations