@fingolfin 2025-02-22
Made it work with Oscar 1.1 (see new code comment)