@SoongNoonien 2025-02-23
> In the meantime: any thoughts on this PR

I'm not so sure if it is worth it. I think this makes the code less nice.